@@ -1,102 +1,102 @@
 import 'package:flutter/material.dart';
 
-TextStyle dtBodyBold({Color color}) => TextStyle(
+TextStyle dtBodyBold({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w700,
     fontSize: 16,
     color: color);
 
-TextStyle dtBoldButtons({Color color}) => TextStyle(
+TextStyle dtBoldButtons({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w700,
     fontSize: 12,
     color: color);
 
-TextStyle dtBodyLarge({Color color}) => TextStyle(
+TextStyle dtBodyLarge({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w700,
     fontSize: 24,
     color: color);
 
-TextStyle dtBodyNormal({Color color}) => TextStyle(
+TextStyle dtBodyNormal({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w400,
     fontSize: 16,
     color: color);
 
-TextStyle dtEyebrowText({Color color}) => TextStyle(
+TextStyle dtEyebrowText({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w600,
     fontSize: 10,
     color: color);
 
-TextStyle dtForCTAButtonText({Color color}) => TextStyle(
+TextStyle dtForCTAButtonText({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w700,
     fontSize: 14,
     color: color);
 
-TextStyle dtFormDefaultText({Color color}) => TextStyle(
+TextStyle dtFormDefaultText({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w600,
     fontSize: 14,
     color: color);
 
-TextStyle dtH3({Color color}) => TextStyle(
+TextStyle dtH3({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w700,
     fontSize: 32,
     color: color);
 
-TextStyle dtMediumBodyText({Color color}) => TextStyle(
+TextStyle dtMediumBodyText({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w400,
     fontSize: 10,
     color: color);
 
-TextStyle dtNunito6008({Color color}) => TextStyle(
+TextStyle dtNunito6008({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w600,
     fontSize: 8,
     color: color);
 
-TextStyle dtNunito60010({Color color}) => TextStyle(
+TextStyle dtNunito60010({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w600,
     fontSize: 10,
     color: color);
 
-TextStyle dtNunito70010({Color color}) => TextStyle(
+TextStyle dtNunito70010({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w700,
     fontSize: 10,
     color: color);
 
-TextStyle dtSmallLabels({Color color}) => TextStyle(
+TextStyle dtSmallLabels({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w800,
     fontSize: 8,
     color: color);
 
-TextStyle dtSmallText({Color color}) => TextStyle(
+TextStyle dtSmallText({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w600,
     fontSize: 12,
     color: color);
 
-TextStyle dtSmallTextBold({Color color}) => TextStyle(
+TextStyle dtSmallTextBold({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w800,
     fontSize: 12,
     color: color);
 
-TextStyle dtTitle({Color color}) => TextStyle(
+TextStyle dtTitle({Color? color}) => TextStyle(
     fontFamily: "NunitoSans",
     fontWeight: FontWeight.w700,
     fontSize: 20,
     color: color);
 
-TextStyle dtCustom({Color color, FontWeight fontWeight, double fontSize}) =>
+TextStyle dtCustom({Color? color, FontWeight? fontWeight, double? fontSize}) =>
     TextStyle(
         fontFamily: "NunitoSans",
         fontWeight: fontWeight,
